% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nas.R
\name{check_na}
\alias{check_na}
\title{checks how many values in a variable in a df are NA}
\usage{
check_na(dataset, variable)
}
\arguments{
\item{dataset}{the dataframe where the variable lives}

\item{variable}{the variable to check}
}
\value{
count of NA values in variable
}
\description{
checks how many values in a variable in a df are NA
}
