% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palettes}
\alias{palettes}
\title{Returns color palettes}
\usage{
palettes(palette)
}
\arguments{
\item{palette}{a palette name}
}
\value{
a color palette of hex codes in vector form
}
\description{
Returns color palettes
}
