% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-examples.R
\name{print_example}
\alias{print_example}
\title{Prints some example code for things I often forget how to do}
\usage{
print_example(x)
}
\arguments{
\item{x}{a string with the example I am looking for}
}
\value{
example code
}
\description{
Prints some example code for things I often forget how to do
}
